Development/feature/ckw

See merge request I528859/huister!30

+++ src/main/java/nl/fontys/s3/huister/business/response/property/GetPropertyDetailResponse.java
@@ -15,5 +15,5 @@ public class GetPropertyDetailResponse {
     private int id,ownerId;
     private String streetName,postCode,description,ownerName,cityName;
     private List<String> imageUrls;
-    private long area,price;
+    private double area,price;
 }
